radiolytics_server/requirements.txt
numpy>=1.21.0
pyaudio>=0.2.11
requests>=2.26.0
flask>=2.0.0
firebase-admin>=5.0.0
python-dotenv>=0.19.0
pydub>=0.25.1
ffmpeg-python>=0.2.0 